rm paranoid
