+++ src/components/Graph/NonSelectableConnection.tsx
@@ -5,7 +5,8 @@ import {MultipointConnection} from '@gravity-ui/graph/react';
  * Наследуется от MultipointConnection и переопределяет поведение
  */
 export class NonSelectableConnection extends MultipointConnection {
-    public override cursor: 'pointer' = 'pointer';
+    //eslint-disable @typescript-eslint/explicit-member-accessibility
+    override cursor: 'pointer' = 'pointer';
 
     // Переопределяем метод для предотвращения выделения при клике
     protected override handleEvent(event: Event) {
