+++ src/utils/prepareQueryExplain.ts
@@ -2,12 +2,11 @@ import type {
     ExplainPlanNodeData,
     GraphNode,
     Link,
+    SimplifiedPlanItem,
     TopologyNodeDataStats,
     TopologyNodeDataStatsItem,
     TopologyNodeDataStatsSection,
-} from '@gravity-ui/paranoid';
-
-import type {SimplifiedPlanItem} from '../store/reducers/query/types';
+} from '../store/reducers/query/types';
 import type {PlanNode, SimplifiedNode} from '../types/api/query';
 
 const CONNECTION_NODE_META_FIELDS = new Set(['PlanNodeId', 'PlanNodeType', 'Node Type', 'Plans']);
